note creation page
