+++ package.json
@@ -14,6 +14,7 @@
     "dexie": "^3.2.2",
     "marked": "^4.2.5",
     "vue": "^3.2.25",
+    "vue-router": "^4.1.6",
     "workbox-precaching": "^6.5.3"
   },
   "devDependencies": {
